# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi User Management Apis Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "usamauddin217@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/permissions": {
            "get": {
                "tags": [
                    "Permissions"
                ],
                "summary": "Get list of Permissions",
                "description": "Returns list of Permissions",
                "operationId": "get permissions",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PermissionResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Permissions"
                ],
                "summary": "Store Single new Permission",
                "description": "Returns Inserted Permission Resource",
                "operationId": "store permission",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PermissionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PermissionResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/permissions/{id}": {
            "get": {
                "tags": [
                    "Permissions"
                ],
                "summary": "Get Permission information",
                "description": "Returns Permission Resource",
                "operationId": "getPermissionById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Permission id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PermissionResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "put": {
                "tags": [
                    "Permissions"
                ],
                "summary": "Update existing Permission",
                "description": "Returns updated Permissions Resource",
                "operationId": "UpdatePermission",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Permission id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PermissionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PermissionResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "PermissionRequest": {
                "title": "Permission Form Request",
                "description": "permission form request body data Validate",
                "required": [
                    "name",
                    "slug"
                ],
                "properties": {
                    "name": {
                        "title": "Permission Name",
                        "description": "Name of the new Permission",
                        "type": "string",
                        "example": "Update Employee"
                    },
                    "slug": {
                        "title": "Permission Slug",
                        "description": "Slug of the new Permission",
                        "type": "string",
                        "example": "update-employee"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "PermissionRequest"
                }
            },
...
            "PermissionResource": {
                "title": "Permission Resource",
                "description": "Permission resource",
                "properties": {
                    "": {
                        "title": "Data",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Permission"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "PermissionResource"
                }
            },
...
            "Permission": {
                "title": "Permission",
                "description": "Permission model",
                "properties": {
                    "permission_id": {
                        "title": "Permission ID",
                        "description": "Permission ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "title": "Permission Name",
                        "description": "Name of the new Permission",
                        "type": "string",
                        "example": "Update Employee"
                    },
                    "slug": {
                        "title": "Permission Slug",
                        "description": "Slug of the new Permission",
                        "type": "string",
                        "example": "update-employee"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Permission"
                }
            },
...
        }
    },
...
}